parameterize set_random_backupstore fixture to test for both S3 and NFS
and move backupstore fixtures to backupstore.py
